Fix bugs in handling of residuals
This fixes three bugs:

1. When a numerical value is used for parameter residual of
applyFilters(), it should (according to manual pages) correspond to
"the cutoff as number of standard deviations estimated with madDiff".
However, instead of multiplying the supplied parameter value with the
calculated standard deviation, the parameter value was used directly.
This is now fixed. When a boolean value was used for the parameter (the
default is TRUE), everything was fine. In practice, the bug only
affected situations with custom-made bin annotations.

2. Same as above, but for highlightFilters().

3. When using the residual filter, applyFilters() used to remove
chromosomes for which the filter was missing altogether (for the
pre-generated bin annotations this meant X and Y). This was changed a
while ago, but highlightFilters() still followed the old behavior. This
is now fixed.